Refactor a bit more
