+++ mistralrs-core/src/pipeline/loaders/mod.rs
@@ -2,9 +2,11 @@ pub(crate) mod auto_device_map;
 mod diffusion_loaders;
 mod normal_loaders;
 mod vision_loaders;
+pub use auto_device_map::AutoDeviceMapParams;
+use auto_device_map::NonMappedSubModel;
 
 use std::{
-    fmt::{self, Debug, Display},
+    fmt::{self, Debug},
     path::PathBuf,
     str::FromStr,
     sync::Arc,
@@ -315,143 +317,6 @@ impl ModelKind {
     }
 }
 
-#[derive(Debug, Clone)]
-pub enum AutoDeviceMapParams {
-    Text {
-        max_seq_len: usize,
-        max_batch_size: usize,
-    },
-    Vision {
-        max_seq_len: usize,
-        max_batch_size: usize,
-        max_image_shape: (usize, usize),
-        max_num_images: usize,
-    },
-}
-
-impl AutoDeviceMapParams {
-    pub fn maybe_promote_to_vision(&self) -> Self {
-        match *self {
-            Self::Text {
-                max_seq_len,
-                max_batch_size,
-            } => Self::Vision {
-                max_seq_len,
-                max_batch_size,
-                max_image_shape: (
-                    Self::DEFAULT_MAX_IMAGE_LENGTH,
-                    Self::DEFAULT_MAX_IMAGE_LENGTH,
-                ),
-                max_num_images: Self::DEFAULT_MAX_NUM_IMAGES,
-            },
-            Self::Vision {
-                max_seq_len,
-                max_batch_size,
-                max_image_shape,
-                max_num_images,
-            } => Self::Vision {
-                max_seq_len,
-                max_batch_size,
-                max_image_shape,
-                max_num_images,
-            },
-        }
-    }
-    pub fn max_seq_len(&self) -> usize {
-        match self {
-            Self::Text { max_seq_len, .. } | Self::Vision { max_seq_len, .. } => *max_seq_len,
-        }
-    }
-}
-
-impl Display for AutoDeviceMapParams {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Text {
-                max_seq_len,
-                max_batch_size,
-            } => write!(
-                f,
-                "text[max_seq_len: {max_seq_len}, max_batch_size: {max_batch_size}]"
-            ),
-            Self::Vision {
-                max_seq_len,
-                max_batch_size,
-                max_image_shape,
-                max_num_images
-            } => write!(
-                f,
-                "vision[max_seq_len: {max_seq_len}, max_batch_size: {max_batch_size}, max_image_shape: {max_image_shape:?}, max_num_images: {max_num_images}]"
-            ),
-        }
-    }
-}
-
-impl AutoDeviceMapParams {
-    pub const DEFAULT_MAX_SEQ_LEN: usize = 4 * 1024;
-    pub const DEFAULT_MAX_BATCH_SIZE: usize = 1;
-    pub const DEFAULT_MAX_NUM_IMAGES: usize = 1;
-    pub const DEFAULT_MAX_IMAGE_LENGTH: usize = 1024;
-
-    pub fn default_text() -> Self {
-        Self::Text {
-            max_seq_len: Self::DEFAULT_MAX_SEQ_LEN,
-            max_batch_size: Self::DEFAULT_MAX_BATCH_SIZE,
-        }
-    }
-
-    pub fn default_vision() -> Self {
-        Self::Vision {
-            max_seq_len: Self::DEFAULT_MAX_SEQ_LEN,
-            max_batch_size: Self::DEFAULT_MAX_BATCH_SIZE,
-            max_num_images: Self::DEFAULT_MAX_NUM_IMAGES,
-            max_image_shape: (
-                Self::DEFAULT_MAX_IMAGE_LENGTH,
-                Self::DEFAULT_MAX_IMAGE_LENGTH,
-            ),
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub(crate) enum NonMappedSubModel {
-    Vision,
-}
-
-impl Display for NonMappedSubModel {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Vision => write!(f, "vision"),
-        }
-    }
-}
-
-fn calculate_key_block_shape(
-    model_config: &dyn ModelConfigLike,
-    dtype: DType,
-    block_size: usize,
-) -> (usize, usize, usize, usize) {
-    let element_size = dtype.size_in_bytes();
-    let x = 16 / element_size;
-    (
-        model_config.num_kv_heads(),
-        model_config.k_head_dim() / x,
-        block_size,
-        x,
-    )
-}
-
-fn calculate_value_block_shape(
-    model_config: &dyn ModelConfigLike,
-    block_size: usize,
-) -> (usize, usize, usize) {
-    (
-        model_config.num_kv_heads(),
-        model_config.v_head_dim(),
-        block_size,
-    )
-}
-
 #[derive(Deserialize)]
 pub struct QuantizationConfigShim {
     quantization_config: Option<QuantizedConfig>,
@@ -519,7 +384,10 @@ pub trait DeviceMappedModelLoader {
         params: &AutoDeviceMapParams,
         prompt_chunksize: usize,
         paged_attn_config: Option<&PagedAttentionConfig>,
-    ) -> Result<DeviceMapMetadata> where Self: Sized {
+    ) -> Result<DeviceMapMetadata>
+    where
+        Self: Sized,
+    {
         auto_device_map::get_device_layers(
             self,
             config,
